backend/requirements.txt
dj-database-url==0.5.0
Django==2.1.7
django-heroku==0.3.1
djangorestframework==3.9.1
djangorestframework-camel-case==1.0.3
djangorestframework-jwt==1.11.0
psycopg2==2.7.7
PyJWT==1.7.1
pytz==2018.9
whitenoise==4.1.2
